fix(ContentView.swift): fix location services call
