+++ FoodBookApp/ContentView.swift
@@ -35,12 +35,6 @@ struct ContentView: View {
                 BrowseView(searchText: $searchText)
                     .tabItem { Label("Browse", systemImage: "magnifyingglass.circle") }
                     .tag(Tabs.browse)
-                    .onAppear{
-                        if locationService.userLocation == nil {
-                            locationService.requestLocationAuthorization()
-                            schedule()
-                        }
-                    }
                 ForYouView()
                     .tabItem { Label("For you", systemImage: "star") }
                     .tag(Tabs.foryou)
@@ -51,32 +45,6 @@ struct ContentView: View {
             }
             .navigationTitle(selectedTab.formattedTitle)
             .modifier(SearchableModifier(isSearchable: selectedTab == .browse, text: $searchText))
-        }
-    }
-}
-
-private func schedule () {
-    let taskId = "Team23.FoodBookApp.contextTask"
-    
-    // Manual Test: e -l objc -- (void)[[BGTaskScheduler sharedScheduler] _simulateLaunchForTaskWithIdentifier:@"Team23.FoodBookApp.contextTask"]
-    BGTaskScheduler.shared.cancel(taskRequestWithIdentifier: taskId)
-    BGTaskScheduler.shared.getPendingTaskRequests { request in
-        print("\(request.count) tasks pending...")
-        
-        // Don't schedule repeated tasks
-        guard request.isEmpty else {
-            return
-        }
-        
-        // Submit a task to be scheduled
-        do {
-            let newTask = BGAppRefreshTaskRequest(identifier: taskId)
-//            newTask.earliestBeginDate = Calendar.current.date(bySettingHour: 12, minute: 10 , second: 0, of: Date())! // Wont execute before said hour.
-            newTask.earliestBeginDate =  Date().addingTimeInterval(120) // Testing - in theory every 2 minutes
-            try BGTaskScheduler.shared.submit(newTask)
-            print("Submitted task...")
-        } catch {
-            print("Could not submit task to be scheduled. \(error.localizedDescription)")
         }
     }
 }
